+++ src/main/java/org/mmarini/wheelly/envs/WorldEnvironment.java
@@ -31,20 +31,19 @@ package org.mmarini.wheelly.envs;
 import com.fasterxml.jackson.databind.JsonNode;
 import org.mmarini.rl.agents.AgentConnector;
 import org.mmarini.rl.envs.*;
-import org.mmarini.wheelly.apis.Complex;
-import org.mmarini.wheelly.apis.RobotCommands;
-import org.mmarini.wheelly.apis.WorldModel;
-import org.mmarini.wheelly.apis.WorldModellerConnector;
+import org.mmarini.wheelly.apis.*;
 import org.mmarini.wheelly.apps.JsonSchemas;
 import org.mmarini.yaml.Locator;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.UnaryOperator;
 
 import static java.lang.Math.round;
+import static java.util.Objects.requireNonNull;
 import static org.mmarini.wheelly.apis.RobotApi.MAX_PPS;
 import static org.mmarini.wheelly.apis.Utils.linear;
 
@@ -52,9 +51,17 @@ import static org.mmarini.wheelly.apis.Utils.linear;
  * Adapts the world model api to RL environment api
  */
 public class WorldEnvironment implements EnvironmentApi {
+    public static final String SCHEMA_NAME = "https://mmarini.org/wheelly/env-world-schema-0.1";
     public static final int MAX_DIRECTION_ACTION = 180;
     public static final int MAX_SENSOR_DIR = 90;
-    public static final String SCHEMA_NAME = "https://mmarini.org/wheelly/env-world-schema-0.1";
+    public static final int NUM_CELL_STATES = 4;
+    public static final int NUM_SECTOR_STATES = 3;
+    public static final int MAX_ROBOT_MAP_DIR = 45;
+    public static final FloatSignalSpec SENSOR_SPEC = new FloatSignalSpec(new long[]{1}, -MAX_SENSOR_DIR, MAX_SENSOR_DIR);
+    public static final FloatSignalSpec ROBOT_MAP_DIR_SPEC = new FloatSignalSpec(new long[]{1}, -MAX_ROBOT_MAP_DIR, MAX_ROBOT_MAP_DIR);
+    public static final int NUM_CAN_MOVE_STATES = 6;
+    public static final IntSignalSpec CAN_MOVE_SPEC = new IntSignalSpec(new long[]{1}, NUM_CAN_MOVE_STATES);
+    public static final int NUM_MARKER_STATE_VALUES = 2;
 
     /**
      * Returns the empty radar from definition
@@ -77,13 +84,41 @@ public class WorldEnvironment implements EnvironmentApi {
         int numSpeeds = locator.path("numSpeeds").getNode(root).asInt();
         int numDirections = locator.path("numDirections").getNode(root).asInt();
         int numSensorDirections = locator.path("numSensorDirections").getNode(root).asInt();
-        return new WorldEnvironment(numSpeeds, numDirections, numSensorDirections);
+        List<String> markerLabels = locator.path("markerLabels").elements(root)
+                .map(l -> l.getNode(root).asText())
+                .toList();
+        return new WorldEnvironment(numSpeeds, numDirections, numSensorDirections, markerLabels);
     }
-
     private final int numSpeeds;
     private final int numDirections;
     private final int numSensorDirections;
     private final Map<String, SignalSpec> actionsSpec;
+
+    /**
+     * Returns the world signal specifications
+     *
+     * @param worldSpec  the world specification
+     * @param numMarkers the number of recognised markers
+     */
+    static Map<String, SignalSpec> createStateSpec(WorldModelSpec worldSpec, long numMarkers) {
+        RobotSpec robotSpec = worldSpec.robotSpec();
+        float maxRadarDistance = (float) robotSpec.maxRadarDistance();
+        int numSectors = worldSpec.numSectors();
+        long radarSize = worldSpec.gridSize();
+        return Map.of(
+                "sensor", SENSOR_SPEC,
+                "robotMapDir", ROBOT_MAP_DIR_SPEC,
+                "distance", new FloatSignalSpec(new long[]{1}, 0, maxRadarDistance),
+                "canMoveStates", CAN_MOVE_SPEC,
+                "sectorStates", new IntSignalSpec(new long[]{numSectors}, NUM_SECTOR_STATES),
+                "sectorDistances", new FloatSignalSpec(new long[]{numSectors}, 0, maxRadarDistance),
+
+                "cellStates", new IntSignalSpec(new long[]{radarSize * radarSize}, NUM_CELL_STATES),
+                "markerStates", new IntSignalSpec(new long[]{numMarkers}, NUM_MARKER_STATE_VALUES),
+                "markerDistances", new FloatSignalSpec(new long[]{numMarkers}, 0, maxRadarDistance),
+                "markerDirections", new FloatSignalSpec(new long[]{numMarkers}, (float) -Math.PI, (float) Math.PI)
+        );
+    }
     private Map<String, Signal> signals0;
     private State prevState;
     private Map<String, Signal> prevActions;
@@ -91,23 +126,7 @@ public class WorldEnvironment implements EnvironmentApi {
     private Consumer<ExecutionResult> onResult;
     private RewardFunction rewardFunc;
     private Map<String, SignalSpec> stateSpec;
-
-    /**
-     * Creates the adapter
-     *
-     * @param numSpeeds           the number of move action speeds
-     * @param numDirections       the number of move action directions
-     * @param numSensorDirections the number of sensor directions
-     */
-    public WorldEnvironment(int numSpeeds, int numDirections, int numSensorDirections) {
-        this.numSpeeds = numSpeeds;
-        this.numDirections = numDirections;
-        this.numSensorDirections = numSensorDirections;
-        this.actionsSpec = Map.of(
-                "move", new IntSignalSpec(new long[]{1}, numDirections * numSpeeds),
-                "sensorAction", new IntSignalSpec(new long[]{1}, numSensorDirections)
-        );
-    }
+    private final List<String> markerLabels;
 
     @Override
     public Map<String, SignalSpec> actionSpec() {
@@ -128,16 +147,36 @@ public class WorldEnvironment implements EnvironmentApi {
         return RobotCommands.haltCommand();
     }
 
-    @Override
-    public void connect(WorldModellerConnector connector) {
-        connector.setOnInference(this::onInference);
-        this.stateSpec = WorldState.stateSpec(connector.worldModelSpec());
+    /**
+     * Creates the adapter
+     *
+     * @param numSpeeds           the number of move action speeds
+     * @param numDirections       the number of move action directions
+     * @param numSensorDirections the number of sensor directions
+     * @param markerLabels        the marker labels
+     */
+    public WorldEnvironment(int numSpeeds, int numDirections, int numSensorDirections, List<String> markerLabels) {
+        this.numSpeeds = numSpeeds;
+        this.numDirections = numDirections;
+        this.numSensorDirections = numSensorDirections;
+        this.markerLabels = requireNonNull(markerLabels);
+        this.actionsSpec = Map.of(
+                "move", new IntSignalSpec(new long[]{1}, numDirections * numSpeeds),
+                "sensorAction", new IntSignalSpec(new long[]{1}, numSensorDirections)
+        );
     }
 
     @Override
     public void connect(AgentConnector agent) {
         onAct = agent::act;
         onResult = agent::observe;
+    }
+
+    @Override
+    public void connect(WorldModellerConnector connector) {
+        connector.setOnInference(this::onInference);
+        WorldModelSpec worldSpec = connector.worldModelSpec();
+        this.stateSpec = createStateSpec(worldSpec, markerLabels.size());
     }
 
     /**
@@ -157,10 +196,10 @@ public class WorldEnvironment implements EnvironmentApi {
      * Returns the halt actions
      */
     public Map<String, Signal> haltActions() {
-        int haltAction = (numSpeeds * (numDirections + 1)) / 2;
+        int haltAction = (numSpeeds * (numDirections + 1)) / NUM_MARKER_STATE_VALUES;
         return Map.of(
                 "move", IntSignal.create(haltAction),
-                "sensorAction", IntSignal.create(numSensorDirections / 2)
+                "sensorAction", IntSignal.create(numSensorDirections / NUM_MARKER_STATE_VALUES)
         );
     }
 
@@ -172,7 +211,7 @@ public class WorldEnvironment implements EnvironmentApi {
     boolean isHalt(Map<String, Signal> actions) {
         int moveAction = actions.get("move").getInt(0);
         // (numSpeeds / 2 + (numDirections * numSpeeds) / 2);
-        int haltAction = (numSpeeds * (numDirections + 1)) / 2;
+        int haltAction = (numSpeeds * (numDirections + 1)) / NUM_MARKER_STATE_VALUES;
         return moveAction == haltAction;
     }
 
@@ -228,11 +267,7 @@ public class WorldEnvironment implements EnvironmentApi {
                 -MAX_SENSOR_DIR, MAX_SENSOR_DIR));
     }
 
-    /**
-     * Sets the reward function
-     *
-     * @param rewardFunc the reward function
-     */
+    @Override
     public void setRewardFunc(RewardFunction rewardFunc) {
         this.rewardFunc = rewardFunc;
     }
@@ -252,11 +287,11 @@ public class WorldEnvironment implements EnvironmentApi {
 
     @Override
     public State state(WorldModel model) {
-        return WorldState.create(model);
+        return WorldState.create(model, stateSpec, markerLabels);
     }
 
     @Override
     public Map<String, SignalSpec> stateSpec() {
-        return this.stateSpec;
+        return stateSpec;
     }
 }